Move all `@connect(http:)` validations to one place
